BackEnd_Flask/requirements.txt
autopep8==1.5.2
cffi==1.14.0
click==7.1.1
cryptography==2.9.2
dnspython==1.16.0
Flask==1.1.2
Flask-Cors==3.0.8
Flask-SocketIO==4.3.0
humanfriendly==8.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.19
protobuf==3.6.1
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
python-engineio==3.13.0
python-socketio==4.6.0
six==1.14.0
validate-email==1.3
Werkzeug==1.0.1
